+++ src/main/java/woowacourse/auth/ui/dto/TokenResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.dto;
+package woowacourse.auth.ui.dto;
 
 public class TokenResponse {
     private String accessToken;
