+++ src/main/java/woowacourse/config/WebConfig.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.config;
+package woowacourse.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpHeaders;
