+++ src/main/java/woowacourse/exception/EmailDuplicateException.java
@@ -1,4 +1,4 @@
-package woowacourse.exception.auth;
+package woowacourse.exception;
 
 public class EmailDuplicateException extends IllegalArgumentException {
     private static final String DEFAULT_MESSAGE = "이미 존재하는 이메일 입니다.";
