+++ src/main/java/woowacourse/exception/InvalidCartItemException.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.exception;
 
 public class InvalidCartItemException extends RuntimeException {
     public InvalidCartItemException() {
