+++ src/main/java/woowacourse/exception/InvalidProductException.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.exception;
 
 public class InvalidProductException extends RuntimeException {
     public InvalidProductException() {
