+++ src/main/java/woowacourse/exception/LoginFailureException.java
@@ -1,4 +1,4 @@
-package woowacourse.exception.auth;
+package woowacourse.exception;
 
 public class LoginFailureException extends IllegalArgumentException {
     private static final String DEFAULT_MESSAGE = "로그인에 실패했습니다";
