+++ src/main/java/woowacourse/exception/NotInCustomerCartItemException.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.exception;
 
 public class NotInCustomerCartItemException extends RuntimeException {
     public NotInCustomerCartItemException() {
