+++ src/main/java/woowacourse/exception/PasswordIncorrectException.java
@@ -1,4 +1,4 @@
-package woowacourse.exception.auth;
+package woowacourse.exception;
 
 public class PasswordIncorrectException extends IllegalArgumentException {
     private static final String DEFAULT_MESSAGE = "비밀번호가 일치하지 않습니다.";
