+++ src/main/java/woowacourse/exception/TokenInvalidException.java
@@ -1,4 +1,4 @@
-package woowacourse.exception.auth;
+package woowacourse.exception;
 
 public class TokenInvalidException extends IllegalArgumentException {
     private static final String DEFAULT_MESSAGE = "유효하지 않은 토큰입니다";
