+++ src/main/java/woowacourse/shoppingcart/service/ProductService.java
@@ -1,11 +1,10 @@
-package woowacourse.shoppingcart.application;
+package woowacourse.shoppingcart.service;
 
+import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import woowacourse.shoppingcart.dao.ProductDao;
 import woowacourse.shoppingcart.domain.Product;
-
-import java.util.List;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
