+++ src/main/java/woowacourse/shoppingcart/ui/CartItemController.java
@@ -1,16 +1,21 @@
 package woowacourse.shoppingcart.ui;
 
+import java.net.URI;
+import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.dto.Request;
-import woowacourse.shoppingcart.application.CartService;
-
-import java.net.URI;
-import java.util.List;
+import woowacourse.shoppingcart.service.CartService;
+import woowacourse.shoppingcart.ui.dto.request.Request;
 
 @RestController
 @RequestMapping("/api/customers/{customerName}/carts")
@@ -28,7 +33,7 @@ public class CartItemController {
 
     @PostMapping
     public ResponseEntity<Void> addCartItem(@Validated(Request.id.class) @RequestBody final Product product,
-                                      @PathVariable final String customerName) {
+                                            @PathVariable final String customerName) {
         final Long cartId = cartService.addCart(product.getId(), customerName);
         final URI responseLocation = ServletUriComponentsBuilder
                 .fromCurrentRequest()
@@ -40,7 +45,7 @@ public class CartItemController {
 
     @DeleteMapping("/{cartId}")
     public ResponseEntity<Void> deleteCartItem(@PathVariable final String customerName,
-                                         @PathVariable final Long cartId) {
+                                               @PathVariable final Long cartId) {
         cartService.deleteCart(customerName, cartId);
         return ResponseEntity.noContent().build();
     }
