+++ src/main/java/woowacourse/shoppingcart/ui/dto/request/Request.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.ui.dto.request;
 
 public interface Request {
     interface id {
