+++ src/test/java/woowacourse/shoppingcart/ui/CustomerControllerTest.java
@@ -10,10 +10,10 @@ import static woowacourse.ShoppingCartFixture.잉_회원생성요청;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import woowacourse.exception.auth.EmailDuplicateException;
-import woowacourse.exception.auth.PasswordIncorrectException;
-import woowacourse.exception.auth.TokenInvalidException;
-import woowacourse.shoppingcart.exception.CustomerNotFoundException;
+import woowacourse.exception.CustomerNotFoundException;
+import woowacourse.exception.EmailDuplicateException;
+import woowacourse.exception.PasswordIncorrectException;
+import woowacourse.exception.TokenInvalidException;
 import woowacourse.shoppingcart.ui.dto.request.CustomerDeleteRequest;
 import woowacourse.shoppingcart.ui.dto.request.CustomerRequest;
 import woowacourse.shoppingcart.ui.dto.request.CustomerUpdatePasswordRequest;
